requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
uvicorn==0.22.0
pandas==2.0.3
python-dotenv==0.21.1
mysql-connector-python==8.1.0
pydantic==2.3.0
jwt==1.3.1